+++ lib/ductwork/models/pipeline.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Ductwork
-  class Pipeline < Ductwork::Record
+  class Pipeline < Ductwork::Record # rubocop:todo Metrics/ClassLength
     has_many :runs,
              class_name: "Ductwork::Run",
              foreign_key: "pipeline_id",
@@ -28,6 +28,7 @@ module Ductwork
     end
 
     class DefinitionError < StandardError; end
+    class ReviveError < StandardError; end
 
     class << self
       attr_reader :pipeline_definition
@@ -130,6 +131,112 @@ module Ductwork
         pipeline_id: id,
         pipeline_klass: klass
       )
+    end
+
+    def revive!(duplicate_context: false)
+      if !halted?
+        raise ReviveError, "Cannot revive #{status} pipeline"
+      end
+
+      last_run = runs.order(started_at: :desc).first
+
+      if last_run.blank?
+        raise ReviveError, "Cannot revive pipeline without previous run"
+      end
+
+      now = Time.current
+      new_run = last_run.dup
+      new_run.triggered_at = now
+      new_run.started_at = now
+      new_run.status = "in_progress"
+
+      Ductwork::Record.transaction do
+        new_run.save!
+        duplicate_successful_branches_and_steps(new_run, last_run, now)
+        duplicate_halted_branches_and_steps(new_run, last_run, now)
+        conditionally_duplicate_context(last_run, new_run, now, duplicate_context)
+        in_progress!
+      end
+
+      self
+    end
+
+    private
+
+    def duplicate_successful_branches_and_steps(new_run, last_run, now)
+      status = %i[advancing waiting completed]
+
+      last_run.branches.where(status:).find_each do |branch|
+        new_branch = branch.dup
+        new_branch.run = new_run
+        new_branch.started_at = now
+        new_branch.completed_at = now
+
+        new_branch.save!
+
+        branch.steps.where(status:).find_each do |step|
+          new_step = step.dup
+          new_step.source_step = step
+          new_step.branch = new_branch
+          new_step.run = new_run
+          new_step.started_at = now
+          new_step.completed_at = now
+
+          new_step.save!
+        end
+      end
+    end
+
+    def duplicate_halted_branches_and_steps(new_run, last_run, now) # rubocop:todo Metrics/AbcSize
+      status = %i[advancing waiting completed]
+
+      last_run.branches.where(status: :halted).find_each do |branch| # rubocop:todo Metrics/BlockLength
+        new_branch = branch.dup
+        new_branch.run = new_run
+        new_branch.status = "in_progress"
+        new_branch.started_at = now
+        new_branch.completed_at = nil
+        new_branch.last_advanced_at = now
+        new_branch.save!
+
+        branch.steps.where(status:).find_each do |step|
+          new_step = step.dup
+          new_step.source_step = step
+          new_step.branch = new_branch
+          new_step.run = new_run
+          new_step.started_at = now
+          new_step.completed_at = now
+
+          new_step.save!
+        end
+
+        failed_step = branch.steps.find_by(status: :failed)
+
+        if failed_step.present?
+          step = new_branch.steps.create!(
+            run: new_run,
+            node: failed_step.node,
+            klass: failed_step.klass,
+            status: :in_progress,
+            to_transition: failed_step.to_transition,
+            started_at: now
+          )
+          Ductwork::Job.enqueue(step)
+        end
+      end
+    end
+
+    def conditionally_duplicate_context(last_run, new_run, now, duplicate_context)
+      if duplicate_context
+        last_run.tuples.find_each do |tuple|
+          new_tuple = tuple.dup
+          new_tuple.run = new_run
+          new_tuple.first_set_at = now
+          new_tuple.last_set_at = now
+
+          new_tuple.save!
+        end
+      end
     end
   end
 end